Fix type error on lfi_x86_bdd signature

+++ src/x64/verify.c
@@ -7,7 +7,7 @@
 #include "lfiv.h"
 #include "fadec.h"
 
-extern int lfi_x86_bdd(uint8_t *);
+extern uint8_t lfi_x86_bdd(uint8_t *);
 
 struct Verifier {
     bool failed;
@@ -228,7 +228,7 @@ bool lfiv_verify_x64(char *code, size_t size, uintptr_t addr, struct LFIVOptions
             insn = &insn_buf[0];
         }
 
-        int n = lfi_x86_bdd(insn);
+        uint8_t n = lfi_x86_bdd(insn);
         if (n == 0) {
             verrmin(&v, "%lx: unknown instruction", v.addr);
             return false;
